Separate aggregation functions in their own modules
